requirements.txt
numpy >= 1.21.1
scikit-learn
matplotlib
pandas
datasets

torch >= 1.9.0
LTNtorch == 0.9
transformers == 4.15.0
